refactor common code, add detailed README
